% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data3.R
\docType{data}
\name{accident_2015.csv.bz2}
\alias{accident_2015.csv.bz2}
\title{Production and farm value of maple products in Canada}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 32166 rows and 52 columns.
}
\source{
National Highway Traffic Safety Administration.
\url{https://www.nhtsa.gov/Data/Fatality-Analysis-Reporting-System-(FARS)}
}
\usage{
accident_2015.csv.bz2
}
\description{
Production and farm value of maple products in Canada
}
\details{
The functions provided for you in this assignment will be using data from
the US National Highway Traffic Safety Administration's Fatality Analysis Reporting
System, which is a nationwide census providing the American public yearly data
regarding fatal injuries suffered in motor vehicle traffic crashes.
}
\keyword{datasets}
