% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{Read the FARS file}
\source{
Data source: US National Highway Traffic Safety Administration's Fatality Analysis Reporting System (FARS),
which is a nationwide census providing the American public yearly data regarding fatal injuries suffered in motor vehicle traffic crashes.
For more information, please see \url{https://crashstats.nhtsa.dot.gov/#/DocumentTypeList/23}.
}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{to enter a database with format csv.}
}
\value{
if file exists, this function read the file and return a database as a data frame. If the extension
is diferent to csv, it can not read the file.
}
\description{
The function \code{fars_read} read a csv file if it exists and forwards the argument a data frame.
}
\note{
To generate a file name, use \code{make_filename}.
}
\examples{
\dontrun{fars_read("accident_2013.csv.bz2")}

}
\seealso{
\code{\link{make_filename}}
}
