% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Read the FARS data with given year}
\source{
Data source: US National Highway Traffic Safety Administration's Fatality Analysis Reporting System (FARS),
which is a nationwide census providing the American public yearly data regarding fatal injuries suffered in motor vehicle traffic crashes.
}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{as a variable to tranform it in an integer value. It will be used
when it generate the name of a file with the last function.}
}
\value{
if the year is in the set: 2013, 2014 or 2015, it will transmute a database in a new data frame depending
on a specific month. Otherwise it will print a warning.
}
\description{
The function save the name of a specific data base according to a year,
read the ddatabase and transmute it drops existing variables as year.
}
\examples{
\dontrun{fars_read_years(2013)}
\dontrun{fars_read_years(c(2013, 2014))}

}
\seealso{
\code{\link{fars_read}}

\code{\link{make_filename}}
}
