% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Summarize FARS data by year}
\source{
Data source: US National Highway Traffic Safety Administration's Fatality Analysis Reporting System (FARS),
which is a nationwide census providing the American public yearly data regarding fatal injuries suffered
in motor vehicle traffic crashes.
}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{as a variable to read the other functions and transmute the data frame.}
}
\value{
a data frame by group of year and month, and summarize by count. It will print the head of the database.
}
\description{
The function transmute the data frame by group, summarize and spread a key-value pair
across the variables year and n.
}
\examples{
\dontrun{fars_summarize_years(2013)}
\dontrun{fars_summarize_years(c(2013, 2014))}

}
