% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{Create a filename}
\usage{
make_filename(year)
}
\arguments{
\item{year}{as a variable to tranform it in an integer value.}
}
\value{
a character vector containing a formatted combination of text and variable value.
}
\description{
The function transform a variable as a integer and print a character vector
containing a formatted combination of text and variable value.
}
\note{
This function will only create a filename, to read in the file, use \code{fars_read}.
}
\examples{
make_filename(2013)

}
\seealso{
\code{\link{fars_read}}
}
